+++ build_test/lib/src/test_builder.dart
@@ -6,48 +6,45 @@ import 'dart:async';
 import 'package:test/test.dart';
 import 'package:build/build.dart';
 import 'package:build_barback/build_barback.dart';
+import 'package:logging/logging.dart';
 
 import 'in_memory_writer.dart';
 import 'in_memory_reader.dart';
 import 'assets.dart';
 
 void _checkOutputs(Map<String, /*String|Matcher*/ dynamic> outputs,
-    [Map<AssetId, DatedString> actualAssets]) {
+    [Map<AssetId, Asset> actualAssets]) {
   if (outputs != null) {
     outputs.forEach((serializedId, contentsMatcher) {
       assert(contentsMatcher is String || contentsMatcher is Matcher);
 
       var assetId = makeAssetId(serializedId);
 
-      /// Check that the writer wrote the assets
-      if (actualAssets != null) {
-        expect(actualAssets, contains(assetId));
-        expect(actualAssets[assetId].value, contentsMatcher,
-            reason: 'Unexpected content for $assetId.');
-      }
-
-      /// Check that the assets exist in [result.outputs].
+      // Check that the asset was produced.
       var actual = actualAssets.remove(assetId);
       expect(actual, isNotNull,
           reason: 'Expected to find $assetId in ${actualAssets}.');
-      expect(actual.value, contentsMatcher,
+      expect(actual.stringContents, contentsMatcher,
           reason: 'Unexpected content for $assetId in result.outputs.');
     });
+    // Check that no extra assets were produced.
+    expect(actualAssets, isEmpty,
+        reason:
+            'Unexpected outputs found `$actualAssets`. Only expected $outputs');
   }
-  expect(actualAssets, isEmpty,
-      reason:
-          'Unexpected outputs found `$actualAssets`. Only expected $outputs');
 }
 
 /// Runs [builder] in a test environment.
 ///
 /// The test environment supplies in-memory build [sourceAssets] to the builders
 /// under test. [outputs] may be optionally provided to verify that the builders
-/// produce the expected output.
+/// produce the expected output. If outputs is omitted the only validation this
+/// method provides is that the build did not `throw`.
 ///
-/// [generateFor] specifies which assets should be given as inputs to the
-/// builder, this can be omitted if every asset in [sourceAssets] should be
-/// considered an input.
+/// [generateFor] or the [isInput] call back can specify which assets should be
+/// given as inputs to the builder. These can be omitted if every asset in
+/// [sourceAssets] should be considered an input. [isInput] precedent over
+/// [generateFor] if both are provided.
 ///
 /// The keys in [sourceAssets] and [outputs] are paths to file assets and the
 /// values are file contents. The paths must use the following format:
@@ -57,26 +54,41 @@ void _checkOutputs(Map<String, /*String|Matcher*/ dynamic> outputs,
 /// Where `PACKAGE_NAME` is the name of the package, and `PATH_WITHIN_PACKAGE`
 /// is the path to a file relative to the package. `PATH_WITHIN_PACKAGE` must
 /// include `lib`, `web`, `bin` or `test`. Example: "myapp|lib/utils.dart".
+///
+/// Callers may optionally provide a [writer] to stub different behavior or do
+/// more complex validation than what is possible with [outputs].
+///
+/// Callers may optionally provide an [onLog] callback to do validaiton on the
+/// logging output of the builder.
 Future testBuilder(Builder builder, Map<String, String> sourceAssets,
     {Set<String> generateFor,
-    Map<String, /*String|Matcher*/ dynamic> outputs}) async {
-  var writer = new InMemoryAssetWriter();
-  final reader = new InMemoryAssetReader(writer.assets);
+    bool isInput(String assetId),
+    String rootPackage,
+    AssetWriter writer,
+    Map<String, /*String|Matcher*/ dynamic> outputs,
+    void onLog(LogRecord log)}) async {
+  writer ??= new InMemoryAssetWriter();
+  final reader = new InMemoryAssetReader();
 
   var inputIds = <AssetId>[];
-  var sourceAssetIds = new Set<AssetId>();
   sourceAssets.forEach((serializedId, contents) {
     var asset = makeAsset(serializedId, contents);
-    writer.writeAsString(asset);
-    sourceAssetIds.add(asset.id);
-    if (generateFor == null || generateFor.contains(serializedId)) {
-      inputIds.add(asset.id);
-    }
+    reader.cacheAsset(asset);
+    inputIds.add(asset.id);
   });
-  await runBuilder(builder, inputIds, reader, writer, const BarbackResolvers());
-  var actualOutputs = <AssetId, DatedString>{}..addAll(writer.assets);
-  for (var assetId in sourceAssetIds) {
-    actualOutputs.remove(assetId);
-  }
+
+  isInput ??= generateFor?.contains ?? (_) => true;
+  inputIds = inputIds.where((id) => isInput('$id'));
+
+  var writerSpy = new AssetWriterSpy(writer);
+  var logger = new Logger('testBuilder');
+  var logSubscription = logger.onRecord.listen(onLog);
+  await runBuilder(
+      builder, inputIds, reader, writerSpy, const BarbackResolvers(),
+      rootPackage: rootPackage, logger: logger);
+  await logSubscription.cancel();
+  var actualOutputs = new Map<AssetId, Asset>.fromIterable(
+      writerSpy.assetsWritten,
+      key: (asset) => asset.id);
   _checkOutputs(outputs, actualOutputs);
 }